requirements.txt
biopython==1.65
coverage==3.7.1
nose==1.3.4
numpy==1.9.1
requests==2.5.1
